fix: support the encrypt auth flow for mcp auth (#9289)
